\name{ActuDistns-package}
\alias{ActuDistns-package}
\alias{ActuDistns}
\docType{package}
\title{
Computes functions for actuarial use
}
\description{
Computes the probability density function, hazard rate function, integrated hazard rate function and the quantile function for 44 commonly used survival models}
\details{
\tabular{ll}{
Package: \tab ActuDistns\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2012-07-12\cr
License: \tab What license is it under?\cr
}
probability density functions, hazard rate functions, integrated hazard rate functions and quantile functions
}
\author{Saralees Nadarajah

Maintainer: Saralees Nadarajah <Saralees.Nadarajah@manchester.ac.uk>
}
\references{
S. Nadarajah, S. A. A. Bakar, Tabulations of survival models for actuarial use, submitted}
\keyword{ package }
